Move DataSourceLiteral to the DataSource annotation
There's a good chance it will be useful for others.
