+++ extensions/agroal/runtime/src/main/java/io/quarkus/agroal/metrics/AgroalGauge.java
@@ -4,6 +4,7 @@ import org.eclipse.microprofile.metrics.Gauge;
 
 import io.agroal.api.AgroalDataSource;
 import io.agroal.api.AgroalDataSourceMetrics;
+import io.quarkus.agroal.DataSource.DataSourceLiteral;
 import io.quarkus.arc.Arc;
 
 public class AgroalGauge implements Gauge<Long> {
