--- extensions/agroal/runtime/src/main/java/io/quarkus/agroal/metrics/DataSourceLiteral.java
@@ -1,20 +1,0 @@
-package io.quarkus.agroal.metrics;
-
-import javax.enterprise.util.AnnotationLiteral;
-
-import io.quarkus.agroal.DataSource;
-
-public class DataSourceLiteral extends AnnotationLiteral<DataSource> implements DataSource {
-
-    private String name;
-
-    public DataSourceLiteral(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String value() {
-        return name;
-    }
-
-}
